+++ src/TicTacToe/TTT.h
@@ -33,9 +33,9 @@ class TicTacToe
     private:
         
         bool _isGameOver = false;
-        int _dimension;
+        byte _dimension;
         Player** board;
-        int _availableIndexesSize;
+        byte _availableIndexesSize;
         INDEX* _availableIndexes;
         bool _xTurn;
 
@@ -50,7 +50,7 @@ class TicTacToe
     public:
         INDEX botMove;
         INDEX humanMove;
-        TicTacToe(int dimension);
+        TicTacToe(byte dimension);
         /// @brief Makes the player move on the specified index
         /// @param index index on the board where the player wants to move
         /// @param type X, O. Is the player X or O
